# Avoid lock contention