@rossabaker 2021-04-19
Techempower is a disservice in both conception and administration, but they continue to be a popular reference.  It's wise to put our best foot forward.  The right thread pool in that context is certainly whatever gets the best top-line ping route numbers, which is all the nuance anybody gets from those damned things.

I look forward to the CE3 advances that make this less contextual.  Again, we don't have the benchmarks to see the timer difference on a fixed pool, but I see no reason this wouldn't also help in that usage.  And most people are using the global pool with Blaze, which is a fork-join pool.  I think this is going to be an improvement for most, if not all, blaze-server users.