@RafalSumislawski 2021-04-19
I think the changes I've proposed in this PR are non-controversial. Less locking should be better. I'll do some more benchmarking with other configuration of threadpools to make sure that I'm not over-optimising for a specific case. I already did some tests with both http4s' `ExecutionContext` and `ContextShift` sharing a `ThreadPoolExecutor`. In that case the baseline and this PR are scoring similarly. The lock in the `ThreadPoolExecutor` being the bottleneck. I'll do some more tests were the thread pools are not shared, and I'll post the result when ready. 

I also have some ideas for reducing the amount of thread shifting. These should be beneficial in their own right (unless they cause some other issues), and should reduce the lock contention in ThreadPoolExecutor.

I'm glad to head about the advances in scheduling in CE3. I'm looking forward to using CE3 in production. 

The way I see TFB in this context is that: http4s itself should be optimised for a wide variety of use-cases for the benefit of it's users. TFB is unable to reproduce this "wide variety", nontheless it can be useful for detecting some performance degrading phenomenon in http4s and quantifying them. As a secondary goal, improving TFB scores is good for PR.